requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
google-cloud-compute==1.14.0
google-cloud-storage==2.10.0
google-cloud-bigquery==3.13.0
google-cloud-iam==2.12.0
google-cloud-monitoring==2.16.0
google-cloud-logging==3.8.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
openai==1.3.7
langchain==0.0.350
langchain-openai==0.0.2
langchain-google-genai==0.0.5
google-generativeai==0.3.2
pydantic==2.5.0
python-dotenv==1.0.0
streamlit==1.28.1
pandas==2.1.3
numpy==1.25.2
plotly==5.17.0
jinja2==3.1.2
aiofiles==23.2.1
asyncio==3.4.3
httpx==0.25.2
redis==5.0.1
celery==5.3.4
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
pytest==7.4.3
pytest-asyncio==0.21.1
click==8.1.7
rich==13.7.0
typer==0.9.0 